chore(git): Initial commit
The idea is to create a webapp that shows a map of the 6th and 7th floor of the
White Chapel building, where GDS currently have their offices.

Ultimately we'd like the following features:

- Has descriptions of each meeting room (e.g. "Meeting room 611 is an 18 person
  open space on the north west side of the 6th floor. The nearest team is Site
  Reliability Engineering")
- Has a nice attractive map
- Works (well enough) with JS disabled
- Works offline
- Is accessible to visually impaired users (via screenreaders / decent contrast
  and font sizes)
- Shows users with geolocation (roughly) where they are on the map
- Can tell people which meeting rooms are booked (if they login with Google and
  have access to the calendar)

The plan is to build out some pure HTML bits first, then gradually add the other
features.

In terms of a rough order:

- HTML description of rooms
- SVG map (moving the HTML bits into the SVG)
- Nice styles
- Check accessibility
- Add service workers so it works offline
- Add geolocation
- Add google login and calendar integration