feat(map): Simple SVG mixed with HTML
It looks like we should be able to get the nice semantic / accessibility
benefits of HTML, while still using SVG to draw the map. Hooray!

The complicated thing that's happening here is that we're interleaving three
markup "languages":

- At the top level we have html5, and do not need to worry about XML compliance
- Once we open the `<svg xmlns="http://www.w3.org/2000/svg">` tag we're in SVG
  world, and we do have to write well formed XML
- Once we're inside the `<foreignObject>` element we're back in HTML land, but
  because we're still inside an `<svg>` we need to make sure our HTML is also
  well-formed XML. Browsers can be very strict about XML well-formedness inside
  SVGs.

For now I've only moved 601 up into the SVG, I'll move the others up in a
separate commit.