feat(map): Absolutely minimal react svg template
The hope is that using react to stamp out the SVG is going to make the whole
thing much more maintainable. At the moment we're doing everything on the server,
but going forward we should be able to load react on the client too and make
the thing "isomorphic / universal".

I'm not using JSX here because I don't feel that the overhead of having a
transpiler is worth it for this project.